third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/json-module/cors-crossorigin-requests-expected.txt

This is a testharness.js-based test.
[FAIL] Imported JSON module with parse error, cross-origin, with CORS
  assert_equals: Unexpected _log value expected "SyntaxError" but got ""
Harness: the test ran to completion.
